DEV-898 рефакторинг истории с svg картинками
