+++ src/components/CollectionSVG/collections.ts
@@ -1,0 +1,736 @@
+import {
+    // 16
+    IconAdmin16,
+    IconArrowDown16,
+    IconArrowDownSolid16,
+    IconArrowLeftSolid16,
+    IconArrowRightSolid16,
+    IconArrowSortingDown16,
+    IconArrowSortingUp16,
+    IconArrowUpSolid16,
+    IconBullet16,
+    IconCalendar16,
+    IconCheckmark16,
+    IconCloseAlt16,
+    IconCompleteFile16,
+    IconDislikeOutline16,
+    IconDrag16,
+    IconEdit16,
+    IconEmptyFile16,
+    IconInfo16,
+    IconLikeOutline16,
+    IconLoading16,
+    IconMinus16,
+    IconModer16,
+    IconModeration16,
+    IconOpenLink16,
+    IconPause16,
+    IconPin16,
+    IconPlus16,
+    IconSearch16,
+    IconStop16,
+    IconVisibilityOff16,
+    IconVisibilityOn16,
+    // 24
+    IconAddCalendar24,
+    IconAlert24,
+    IconArrowDown24,
+    IconArrowDownSolid24,
+    IconArrowLeft24,
+    IconArrowLeftSolid24,
+    IconArrowNW24,
+    IconArrowRight24,
+    IconArrowRightSolid24,
+    IconArrowTale24,
+    IconArrowTaleCurve24,
+    IconArrowUp24,
+    IconArrowUpSolid24,
+    IconAttendees24,
+    IconBell24,
+    IconBookmark24,
+    IconBookmarkSolid24,
+    IconCalendar24,
+    IconCard24,
+    IconCheck24,
+    IconCheckCircle24,
+    IconClose24,
+    IconCloseCircle24,
+    IconCopy24,
+    IconDelete24,
+    IconDocument24,
+    IconDrag24,
+    IconEdit24,
+    IconEnter24,
+    IconEnterFullscreen24,
+    IconExitFullscreen24,
+    IconHand24,
+    IconHome24,
+    IconInfo24,
+    IconLink24,
+    IconLoading24,
+    IconLock24,
+    IconLogin24,
+    IconLogout24,
+    IconMail24,
+    IconMenu24,
+    IconMenuLeft24,
+    IconMenuRight24,
+    IconMinus24,
+    IconMinusCircle24,
+    IconMore24,
+    IconMessage24,
+    IconPhoto24,
+    IconPin24,
+    IconPlay24,
+    IconPlaySolid24,
+    IconPlus24,
+    IconPlusCircle24,
+    IconProfile24,
+    IconQrcode24,
+    IconQuestion24,
+    IconRedo24,
+    IconReload24,
+    IconRouble24,
+    IconSale24,
+    IconSearch24,
+    IconServices24,
+    IconSettings24,
+    IconShare24,
+    IconSocket24,
+    IconSwapHor24,
+    IconThumbDown24,
+    IconThumbDownSolid24,
+    IconThumbUp24,
+    IconThumbUpSolid24,
+    IconTicket24,
+    IconUndo24,
+    IconUnlock24,
+    IconVisibilityOff24,
+    IconVisibilityOn24,
+    IconWarning24,
+    IconWebsite24,
+    IconWidget24,
+    // map
+    Pin64,
+    // vendor
+    // appShops
+    IconPlayVideo,
+    IconAppStore,
+    // bank
+    IconGooglePlay,
+    IconBankAlfa,
+    IconBankAlfaWhite,
+    IconBankGazprom,
+    IconBankGazpromWhite,
+    IconBankOtkritie,
+    IconBankOtkritieWhite,
+    IconBankPochta,
+    IconBankPromsvyaz,
+    IconBankPromsvyazWhite,
+    IconBankRaiffeisen,
+    IconBankRosbank,
+    IconBankSber,
+    IconBankSberWhite,
+    IconBankTinkoff,
+    IconBankVTB,
+    // logo
+    Logo,
+    SumsubLogo,
+    // paymentSystems
+    IconApplePay,
+    IconQiwi,
+    IconWebmoney,
+    IconYandexMoney,
+    // socialMedia
+    IconFacebook,
+    IconInstagram,
+    IconTelegram,
+    IconTwitter,
+    IconVk,
+    IconYoutube,
+} from '../../icons';
+import {ICollectionSVGProps} from './CollectionTable';
+
+export const collectionSVG16: ICollectionSVGProps[] = [
+    {
+        SVGSrc: IconAdmin16,
+        title: 'icon-admin-16',
+    },
+    {
+        SVGSrc: IconArrowDown16,
+        title: 'icon-arrow-down-16',
+    },
+    {
+        SVGSrc: IconArrowDownSolid16,
+        title: 'icon-arrow-down-solid-16',
+    },
+    {
+        SVGSrc: IconArrowLeftSolid16,
+        title: 'icon-arrow-left-solid-16',
+    },
+    {
+        SVGSrc: IconArrowRightSolid16,
+        title: 'icon-arrow-right-solid-16',
+    },
+    {
+        SVGSrc: IconArrowSortingDown16,
+        title: 'icon-arrow-sorting-down-16',
+    },
+    {
+        SVGSrc: IconArrowSortingUp16,
+        title: 'icon-arrow-sorting-up-16',
+    },
+    {
+        SVGSrc: IconArrowUpSolid16,
+        title: 'icon-arrow-up-solid-16',
+    },
+    {
+        SVGSrc: IconBullet16,
+        title: 'icon-bullet-16',
+    },
+    {
+        SVGSrc: IconCalendar16,
+        title: 'icon-calendar-16',
+    },
+    {
+        SVGSrc: IconCheckmark16,
+        title: 'icon-checkmark-16',
+    },
+    {
+        SVGSrc: IconCloseAlt16,
+        title: 'icon-close-alt-16',
+    },
+    {
+        SVGSrc: IconCompleteFile16,
+        title: 'icon-complete-file-16',
+    },
+    {
+        SVGSrc: IconDislikeOutline16,
+        title: 'icon-dislike-outline-16',
+    },
+    {
+        SVGSrc: IconDrag16,
+        title: 'icon-drag-16',
+    },
+    {
+        SVGSrc: IconEdit16,
+        title: 'icon-edit-16',
+    },
+    {
+        SVGSrc: IconEmptyFile16,
+        title: 'icon-empty-file-16',
+    },
+    {
+        SVGSrc: IconInfo16,
+        title: 'icon-info-16',
+    },
+    {
+        SVGSrc: IconLikeOutline16,
+        title: 'icon-like-outline-16',
+    },
+    {
+        SVGSrc: IconLoading16,
+        title: 'icon-loading-16',
+    },
+    {
+        SVGSrc: IconMinus16,
+        title: 'icon-minus-16',
+    },
+    {
+        SVGSrc: IconModer16,
+        title: 'icon-moder-16',
+    },
+    {
+        SVGSrc: IconModeration16,
+        title: 'icon-moderation-16',
+    },
+    {
+        SVGSrc: IconOpenLink16,
+        title: 'icon-open-link-16',
+    },
+    {
+        SVGSrc: IconPause16,
+        title: 'icon-pause-16',
+    },
+    {
+        SVGSrc: IconPin16,
+        title: 'icon-pin-16',
+    },
+    {
+        SVGSrc: IconPlus16,
+        title: 'icon-plus-16',
+    },
+    {
+        SVGSrc: IconSearch16,
+        title: 'icon-search-16',
+    },
+    {
+        SVGSrc: IconStop16,
+        title: 'icon-stop-16',
+    },
+    {
+        SVGSrc: IconVisibilityOff16,
+        title: 'icon-visibility-off-16',
+    },
+    {
+        SVGSrc: IconVisibilityOn16,
+        title: 'icon-visibility-on-16',
+    },
+];
+
+export const collectionSVG24: ICollectionSVGProps[] = [
+    {
+        SVGSrc: IconAddCalendar24,
+        title: 'icon-add-calendar-24',
+    },
+    {
+        SVGSrc: IconAlert24,
+        title: 'icon-alert-24',
+    },
+    {
+        SVGSrc: IconArrowDown24,
+        title: 'icon-arrow-down-24',
+    },
+    {
+        SVGSrc: IconArrowDownSolid24,
+        title: 'icon-arrow-down-solid-24',
+    },
+    {
+        SVGSrc: IconArrowLeft24,
+        title: 'icon-arrow-left-24',
+    },
+    {
+        SVGSrc: IconArrowLeftSolid24,
+        title: 'icon-arrow-left-solid-24',
+    },
+    {
+        SVGSrc: IconArrowNW24,
+        title: 'icon-arrow-nw-24',
+    },
+    {
+        SVGSrc: IconArrowRight24,
+        title: 'icon-arrow-right-24',
+    },
+    {
+        SVGSrc: IconArrowRightSolid24,
+        title: 'icon-arrow-right-solid-24',
+    },
+    {
+        SVGSrc: IconArrowTale24,
+        title: 'icon-arrow-tale-24',
+    },
+    {
+        SVGSrc: IconArrowTaleCurve24,
+        title: 'icon-arrow-tale-curve-24',
+    },
+    {
+        SVGSrc: IconArrowUp24,
+        title: 'icon-arrow-up-24',
+    },
+    {
+        SVGSrc: IconArrowUpSolid24,
+        title: 'icon-arrow-up-solid-24',
+    },
+    {
+        SVGSrc: IconAttendees24,
+        title: 'icon-attendees-24',
+    },
+    {
+        SVGSrc: IconBell24,
+        title: 'icon-bell-24',
+    },
+    {
+        SVGSrc: IconBookmark24,
+        title: 'icon-bookmark-24',
+    },
+    {
+        SVGSrc: IconBookmarkSolid24,
+        title: 'icon-bookmark-solid-24',
+    },
+    {
+        SVGSrc: IconCalendar24,
+        title: 'icon-calendar-24',
+    },
+    {
+        SVGSrc: IconCard24,
+        title: 'icon-card-24',
+    },
+    {
+        SVGSrc: IconCheck24,
+        title: 'icon-check-24',
+    },
+    {
+        SVGSrc: IconCheckCircle24,
+        title: 'icon-check-circle-24',
+    },
+    {
+        SVGSrc: IconClose24,
+        title: 'icon-close-24',
+    },
+    {
+        SVGSrc: IconCloseCircle24,
+        title: 'icon-close-circle-24',
+    },
+    {
+        SVGSrc: IconCopy24,
+        title: 'icon-copy-24',
+    },
+    {
+        SVGSrc: IconDelete24,
+        title: 'icon-delete-24',
+    },
+    {
+        SVGSrc: IconDocument24,
+        title: 'icon-document-24',
+    },
+    {
+        SVGSrc: IconDrag24,
+        title: 'icon-drag-24',
+    },
+    {
+        SVGSrc: IconEdit24,
+        title: 'icon-edit-24',
+    },
+    {
+        SVGSrc: IconEnter24,
+        title: 'icon-enter-24',
+    },
+    {
+        SVGSrc: IconEnterFullscreen24,
+        title: 'icon-enter-fullscreen-24',
+    },
+    {
+        SVGSrc: IconExitFullscreen24,
+        title: 'icon-exit-fullscreen-24',
+    },
+    {
+        SVGSrc: IconHand24,
+        title: 'icon-hand-24',
+    },
+    {
+        SVGSrc: IconHome24,
+        title: 'icon-home-24',
+    },
+    {
+        SVGSrc: IconInfo24,
+        title: 'icon-info-24',
+    },
+    {
+        SVGSrc: IconLink24,
+        title: 'icon-link-24',
+    },
+    {
+        SVGSrc: IconLoading24,
+        title: 'icon-loading-24',
+    },
+    {
+        SVGSrc: IconLock24,
+        title: 'icon-lock-24',
+    },
+    {
+        SVGSrc: IconLogin24,
+        title: 'icon-login-24',
+    },
+    {
+        SVGSrc: IconLogout24,
+        title: 'icon-logout-24',
+    },
+    {
+        SVGSrc: IconMail24,
+        title: 'icon-mail-24',
+    },
+    {
+        SVGSrc: IconMenu24,
+        title: 'icon-menu-24',
+    },
+    {
+        SVGSrc: IconMenuLeft24,
+        title: 'icon-menu-left-24',
+    },
+    {
+        SVGSrc: IconMenuRight24,
+        title: 'icon-menu-right-24',
+    },
+    {
+        SVGSrc: IconMinus24,
+        title: 'icon-minus-24',
+    },
+    {
+        SVGSrc: IconMinusCircle24,
+        title: 'icon-minus-circle-24',
+    },
+    {
+        SVGSrc: IconMore24,
+        title: 'icon-more-24',
+    },
+    {
+        SVGSrc: IconMessage24,
+        title: 'icon-message-24',
+    },
+    {
+        SVGSrc: IconPhoto24,
+        title: 'icon-photo-24',
+    },
+    {
+        SVGSrc: IconPin24,
+        title: 'icon-pin-24',
+    },
+    {
+        SVGSrc: IconPlay24,
+        title: 'icon-play-24',
+    },
+    {
+        SVGSrc: IconPlaySolid24,
+        title: 'icon-play-solid-24',
+    },
+    {
+        SVGSrc: IconPlus24,
+        title: 'icon-plus-24',
+    },
+    {
+        SVGSrc: IconPlusCircle24,
+        title: 'icon-plus-circle-24',
+    },
+    {
+        SVGSrc: IconProfile24,
+        title: 'icon-profile-24',
+    },
+    {
+        SVGSrc: IconQrcode24,
+        title: 'icon-qrcode-24',
+    },
+    {
+        SVGSrc: IconQuestion24,
+        title: 'icon-question-24',
+    },
+    {
+        SVGSrc: IconRedo24,
+        title: 'icon-redo-24',
+    },
+    {
+        SVGSrc: IconReload24,
+        title: 'icon-reload-24',
+    },
+    {
+        SVGSrc: IconRouble24,
+        title: 'icon-rouble-24',
+    },
+    {
+        SVGSrc: IconSale24,
+        title: 'icon-sale-24',
+    },
+    {
+        SVGSrc: IconSearch24,
+        title: 'icon-search-24',
+    },
+    {
+        SVGSrc: IconServices24,
+        title: 'icon-services-24',
+    },
+    {
+        SVGSrc: IconSettings24,
+        title: 'icon-settings-24',
+    },
+    {
+        SVGSrc: IconShare24,
+        title: 'icon-share-24',
+    },
+    {
+        SVGSrc: IconSocket24,
+        title: 'icon-socket-24',
+    },
+    {
+        SVGSrc: IconSwapHor24,
+        title: 'icon-swap-hor-24',
+    },
+    {
+        SVGSrc: IconThumbDown24,
+        title: 'icon-thumb-down-24',
+    },
+    {
+        SVGSrc: IconThumbDownSolid24,
+        title: 'icon-thumb-down-solid-24',
+    },
+    {
+        SVGSrc: IconThumbUp24,
+        title: 'icon-thumb-up-24',
+    },
+    {
+        SVGSrc: IconThumbUpSolid24,
+        title: 'icon-thumb-up-solid-24',
+    },
+    {
+        SVGSrc: IconTicket24,
+        title: 'icon-ticket-24',
+    },
+    {
+        SVGSrc: IconUndo24,
+        title: 'icon-undo-24',
+    },
+    {
+        SVGSrc: IconUnlock24,
+        title: 'icon-unlock-24',
+    },
+    {
+        SVGSrc: IconVisibilityOff24,
+        title: 'icon-visibility-off-24',
+    },
+    {
+        SVGSrc: IconVisibilityOn24,
+        title: 'icon-visibility-on-24',
+    },
+    {
+        SVGSrc: IconWarning24,
+        title: 'icon-warning-24',
+    },
+    {
+        SVGSrc: IconWebsite24,
+        title: 'icon-website-24',
+    },
+    {
+        SVGSrc: IconWidget24,
+        title: 'icon-widget-24',
+    },
+];
+
+export const collectionSVGMap: ICollectionSVGProps[] = [
+    {
+        SVGSrc: Pin64,
+        title: 'pin-64',
+    },
+];
+
+export const collectionSVGAppShops: ICollectionSVGProps[] = [
+    {
+        SVGSrc: IconPlayVideo,
+        title: 'icon-play-video',
+    },
+    {
+        SVGSrc: IconAppStore,
+        title: 'icon-app-store',
+    },
+];
+
+export const collectionSVGAppbank: ICollectionSVGProps[] = [
+    {
+        SVGSrc: IconGooglePlay,
+        title: 'icon-google-play',
+    },
+    {
+        SVGSrc: IconBankAlfa,
+        title: 'icon-bank-alfa',
+    },
+    {
+        SVGSrc: IconBankAlfaWhite,
+        title: 'icon-bank-alfa-white',
+    },
+    {
+        SVGSrc: IconBankGazprom,
+        title: 'icon-bank-gazprom',
+    },
+    {
+        SVGSrc: IconBankGazpromWhite,
+        title: 'icon-bank-gazprom-white',
+    },
+    {
+        SVGSrc: IconBankOtkritie,
+        title: 'icon-bank-otkritie',
+    },
+    {
+        SVGSrc: IconBankOtkritieWhite,
+        title: 'icon-bank-otkritie-white',
+    },
+    {
+        SVGSrc: IconBankPochta,
+        title: 'icon-bank-pochta',
+    },
+    {
+        SVGSrc: IconBankPromsvyaz,
+        title: 'icon-bank-promsvyaz',
+    },
+    {
+        SVGSrc: IconBankPromsvyazWhite,
+        title: 'icon-bank-promsvyaz-white',
+    },
+    {
+        SVGSrc: IconBankRaiffeisen,
+        title: 'icon-bank-raiffeisen',
+    },
+    {
+        SVGSrc: IconBankRosbank,
+        title: 'icon-bank-rosbank',
+    },
+    {
+        SVGSrc: IconBankSber,
+        title: 'icon-bank-sber',
+    },
+    {
+        SVGSrc: IconBankSberWhite,
+        title: 'icon-bank-sber-white',
+    },
+    {
+        SVGSrc: IconBankTinkoff,
+        title: 'icon-bank-tinkoff',
+    },
+    {
+        SVGSrc: IconBankVTB,
+        title: 'icon-bank-vtb',
+    },
+];
+
+export const collectionSVGLogo: ICollectionSVGProps[] = [
+    {
+        SVGSrc: Logo,
+        title: 'logo',
+    },
+    {
+        SVGSrc: SumsubLogo,
+        title: 'sumsub-logo',
+    },
+];
+
+export const collectionSVGPaymentSystems: ICollectionSVGProps[] = [
+    {
+        SVGSrc: IconApplePay,
+        title: 'icon-apple-pay',
+    },
+    {
+        SVGSrc: IconQiwi,
+        title: 'icon-qiwi',
+    },
+    {
+        SVGSrc: IconWebmoney,
+        title: 'icon-webmoney',
+    },
+    {
+        SVGSrc: IconYandexMoney,
+        title: 'icon-yandex-money',
+    },
+];
+
+export const collectionSVGSocialMedia: ICollectionSVGProps[] = [
+    {
+        SVGSrc: IconFacebook,
+        title: 'icon-facebook',
+    },
+    {
+        SVGSrc: IconInstagram,
+        title: 'icon-instagram',
+    },
+    {
+        SVGSrc: IconTelegram,
+        title: 'icon-telegram',
+    },
+    {
+        SVGSrc: IconTwitter,
+        title: 'icon-twitter',
+    },
+    {
+        SVGSrc: IconVk,
+        title: 'icon-vk',
+    },
+    {
+        SVGSrc: IconYoutube,
+        title: 'icon-youtube',
+    },
+];
